requirements.txt
###### Requirements without Version Specifiers ######
ml-collections
av
timm
wandb
pytorchvideo
h5py
MultiScaleDeformableAttention

###### Requirements with Version Specifiers ######